requirements.txt
websockets==12.0
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.0
requests==2.31.0
opuslib==3.0.1 